fixed the interrupts for later rust versions

+++ src/gdt.rs
@@ -24,8 +24,8 @@ lazy_static! {
     static ref TSS: TaskStateSegment = {
         let mut tss = TaskStateSegment::new();
         tss.interrupt_stack_table[DOUBLE_FAULT_IST_INDEX as usize] = {
-            const STACK_SIZE: usize = 4096 * 5;
-            static mut STACK: [u8; STACK_SIZE] = [0; STACK_SIZE];
+            const STACK_SIZE: u64 = 4096 * 5;
+            static mut STACK: [u8; STACK_SIZE as usize] = [0; STACK_SIZE as usize];
 
             let stack_start = VirtAddr::from_ptr(&raw const STACK);
             let stack_end = stack_start + STACK_SIZE;
